+++ README.md
@@ -12,7 +12,43 @@
 
 ## Desafio
 
+<center>
+Como podemos ver aqui (linkar com a página do Trabalho Informal), o trabalho informal é um assunto complexo que afeta não só as pessoas, mas também a sociedade como um todo.
+
+O nosso projeto tem como objetivo ajudar pessoas em situação de trabalho informal a se prepararem para algo que consequentemente chegará para todos: o Futuro! 
+Então, como uma forma de conscientização financeira, preparamos um material para auxiliá-las a se prepararem para o futuro, pensando principalmente no risco que elas correm em situações de inatividade: seja doença, perda de renda ou aposentadoria.
+
+Para isso vamos mostrar: (Linkar com suas devidas páginas)
+- Possíveis investimentos, e diferenças
+- Pessoas que podem auxiliá-las no processo
+- Principais corretoras/bancos do mercado
+- Simulador de investimentos
+
+O objetivo do projeto não é indicar nenhum tipo de produto ou empresa, mas sim conscientizar pessoas que trabalham de forma informal que elas precisam se preparar para o futuro e que existem diversas maneiras/empresas/pessoas que podem ajuda-las. 
+</center>
+
 ## Solução
+
+- [x] Utilização de fluxo de trabalho: [git flow](https://www.atlassian.com/br/git/tutorials/comparing-workflows/gitflow-workflow)
+- [x] Construção do Banco de Dados MySQL
+- [x] Instalação das libs:
+  - [x] [express](http://expressjs.com/pt-br/)
+  - [x] [nodemon](https://nodemon.io/)
+  - [x] [ORM sequelize](https://sequelize.org/)
+  - [x] [mysql2](https://www.npmjs.com/package/mysql2)
+- [x] Criação dos endpoints:
+  - [x] GET`/`
+  - [x] GET - `/simulacao`
+  - [x] GET - `/simulacao/:id`
+  - [x] POST `/simulacao`
+  - [x] DELETE `/simulacao/:id`
+  - [x] PUT `/simulacao/:id`
+  
+  - [x] GET - `/usuario`
+  - [x] GET - `/usuario/:id`
+  - [x] POST `/usuario`
+  - [x] DELETE `/usuario/:id`
+  - [x] PUT `/usuario/:id`
 
 ## Projeto Finalizado
 
@@ -25,3 +61,11 @@
 <a target="_blank" href="https://icons8.com/icon/108784/javascript">JavaScript icon by Icons8</a>
 <a target="_blank" href="https://icons8.com/icon/20906/git">Git icon by Icons8</a>
 <a target="_blank" href="https://icons8.com/icon/hsPbhkOH4FMe/node-js">Node Js icon by Icons8</a>
+
+- [ ] front - (estático mesmo)
+- [ ] conecta front e back para exibir gráfico
+- [ ] script para conectar o banco (heroku?)
+- [ ] subir a api (heroku?)
+- [ ] documentar api
+- [ ] componentizar
+- [ ] autenticação
