+++ teste-chart.html
@@ -45,7 +45,7 @@
         </p>
     </footer>
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-    <script type="text/javascript" src="script/index.js"></script>
+    <script type="text/javascript" src="script/chart.js.js"></script>
 </body>
 
 </html>
